unify inputs for creating a project
